Quality of life Refactoring. Updated README
