+++ package.json
@@ -46,6 +46,7 @@
     ]
   },
   "devDependencies": {
-    "dotenv": "^16.0.3"
+    "dotenv": "^16.0.3",
+    "prettier": "^2.8.8"
   }
 }
