+++ src/App.tsx
@@ -1,16 +1,16 @@
 import React, { useEffect, useState } from 'react';
 import { BrowserRouter as Router, Route, Routes, useLocation } from 'react-router-dom';
 import './App.css';
-import HomePage from './pages/HomePage';
-import RoomPage from './pages/RoomPage';
+import HomePage from './pages/HomePage/HomePage';
+import RoomPage from './pages/RoomPage/RoomPage';
 
 function App() {
   return (
     <div className="App">
       <Router>
         <Routes>
-          <Route path="/" element={<HomePage/>} />
-          <Route path="/room" element={<RoomPage/>} />
+          <Route path="/" element={<HomePage />} />
+          <Route path="/room" element={<RoomPage />} />
         </Routes>
       </Router>
     </div>
