+++ src/components/BidOption/BidOption.module.css
