+++ src/components/Bids/Bids.module.css
