+++ src/components/GameInfo/GameInfo.module.css
