+++ src/components/PlayerCard/PlayerCard.tsx
@@ -7,15 +7,20 @@ interface PlayerCardProps {
   card: string;
   suit: string;
   valid: boolean;
-  handleCardClick: (card: string) => void
+  handleCardClick: (card: string) => void;
 }
 
 // Define the PlayerHand component
 const PlayerCard: React.FC<PlayerCardProps> = ({ card, suit, valid, handleCardClick }) => {
-
-  const validClass = valid ? styles.valid : ''
+  const validClass = valid ? styles.valid : '';
   return (
-    <div className={`${styles.card} ${suit} ${validClass}`} style={{backgroundImage: 'url("/assets/images/png-cards/'+card+'.png")'}} onClick={() => { if(valid) handleCardClick(card)}}></div>
+    <div
+      className={`${styles.card} ${suit} ${validClass}`}
+      style={{ backgroundImage: 'url("/assets/images/png-cards/' + card + '.png")' }}
+      onClick={() => {
+        if (valid) handleCardClick(card);
+      }}
+    ></div>
   );
 };
