+++ src/components/TableCard/TableCard.tsx
@@ -11,13 +11,15 @@ interface TableCardProps {
 
 // Define the PlayerHand component
 const TableCard: React.FC<TableCardProps> = ({ card, seatNumber, playerSeat }) => {
-
-    let styles_arr = [styles.seat_1_card,styles.seat_2_card,styles.seat_3_card,styles.seat_4_card]
-    let offset = playerSeat-1
-    let index = (seatNumber - offset) <= 0  ? (seatNumber - offset) + 4 : seatNumber - offset
+  let styles_arr = [styles.seat_1_card, styles.seat_2_card, styles.seat_3_card, styles.seat_4_card];
+  let offset = playerSeat - 1;
+  let index = seatNumber - offset <= 0 ? seatNumber - offset + 4 : seatNumber - offset;
 
   return (
-    <div className={`${styles.card} ${styles_arr[index-1]}`}  style={{backgroundImage: 'url("/assets/images/png-cards/'+card+'.png")'}}></div>
+    <div
+      className={`${styles.card} ${styles_arr[index - 1]}`}
+      style={{ backgroundImage: 'url("/assets/images/png-cards/' + card + '.png")' }}
+    ></div>
   );
 };
