+++ src/components/TableCards/TableCards.tsx
@@ -1,7 +1,8 @@
 // Import React and any necessary dependencies
 import React from 'react';
 import styles from './TableCards.module.css';
-import TableCard from './TableCard';
+import TableCard from '../TableCard/TableCard';
+import { CARD_SUIT_CLUBS, CARD_SUIT_DIAMONDS, CARD_SUIT_HEARTS, CARD_SUIT_SPADES } from '../../utils/Constant';
 
 // Define a type for the component's props (if needed)
 interface TableCardsProps {
@@ -24,16 +25,14 @@ interface TableCardsProps {
 
 // Define the PlayerHand component
 const TableCards: React.FC<TableCardsProps> = ({ playeTurn, playerTurnName, playerName, playersCards, playerSeat }) => {
-
-
-    const text = playerName == playerTurnName ? 'Your turn' : `${playerTurnName}'s turn`
+  const text = playerName == playerTurnName ? 'Your turn' : `${playerTurnName}'s turn`;
   return (
     <div className={styles.table}>
-        <div className={styles.player_turn}>{text}</div>
-        <TableCard card={playersCards[0]} seatNumber={1} playerSeat={playerSeat}/>
-        <TableCard card={playersCards[1]} seatNumber={2} playerSeat={playerSeat}/>
-        <TableCard card={playersCards[2]} seatNumber={3} playerSeat={playerSeat}/>
-        <TableCard card={playersCards[3]} seatNumber={4} playerSeat={playerSeat}/>
+      <div className={styles.player_turn}>{text}</div>
+      <TableCard card={playersCards[0]} seatNumber={1} playerSeat={playerSeat} />
+      <TableCard card={playersCards[1]} seatNumber={2} playerSeat={playerSeat} />
+      <TableCard card={playersCards[2]} seatNumber={3} playerSeat={playerSeat} />
+      <TableCard card={playersCards[3]} seatNumber={4} playerSeat={playerSeat} />
     </div>
   );
 };
