+++ src/utils/Constant.ts
@@ -1,0 +1,5 @@
+// Cards Suits
+export const CARD_SUIT_HEARTS: string = "hearts";
+export const CARD_SUIT_SPADES: string = "spades";
+export const CARD_SUIT_CLUBS: string = "clubs";
+export const CARD_SUIT_DIAMONDS: string = "diamonds";
